+++ honei/backends/cuda/up_vel_dir_grid.cu
@@ -28,7 +28,7 @@ namespace honei
                 unsigned long * types,
                 unsigned long size)
         {
-            int idx = (blockDim.y * blockIdx.y * gridDim.x * blockDim.x) + (blockDim.x * blockIdx.x) + threadIdx.x;
+            unsigned long idx = (blockDim.y * blockIdx.y * gridDim.x * blockDim.x) + (blockDim.x * blockIdx.x) + threadIdx.x;
             if (idx < size)
             {
                 unsigned long i(idx);
